+++ Controllers/TaskToDoController.cs
@@ -18,13 +18,13 @@ namespace ToDoListAPI.Controllers
         [HttpGet]
         public async Task<ActionResult<List<TaskToDo>>> GetTasksToDo()
         {
-            return Ok(await _taskToDoService.GetTasksToDo());
+            return Ok(await _taskToDoService.TaskToDoRepository.GetTasksToDo());
         }
 
         [HttpGet("{id:int}")]
         public async Task<ActionResult<TaskToDo>> GetTaskToDo(int id)
         {
-            var result = await _taskToDoService.GetTaskToDo(id);
+            var result = await _taskToDoService.TaskToDoRepository.GetTaskToDo(id);
 
             if (result == null)
             {
@@ -42,7 +42,7 @@ namespace ToDoListAPI.Controllers
                 return BadRequest();
             }
 
-            var createdTaskToDo = await _taskToDoService.AddTaskToDo(task);
+            var createdTaskToDo = await _taskToDoService.TaskToDoRepository.AddTaskToDo(task);
 
             return CreatedAtAction(nameof(GetTaskToDo),
                 new { Id = createdTaskToDo.Id}, createdTaskToDo);
@@ -56,27 +56,27 @@ namespace ToDoListAPI.Controllers
                 return BadRequest("Task Id mismatch");
             }
 
-            var taskToDoToUpdate = await _taskToDoService.GetTaskToDo(request.Id);
+            var taskToDoToUpdate = await _taskToDoService.TaskToDoRepository.GetTaskToDo(request.Id);
 
             if (taskToDoToUpdate == null)
             {
                 return NotFound($"TaskToDo with Id = {request.Id} not found");
             }
 
-            return await _taskToDoService.UpdateTaskToDo(request);
+            return await _taskToDoService.TaskToDoRepository.UpdateTaskToDo(request);
         }
 
         [HttpDelete("{id}")]
         public async Task<ActionResult<TaskToDo?>> Delete(int id)
         {
-            var taskToDoToDelete = await _taskToDoService.GetTaskToDo(id);
+            var taskToDoToDelete = await _taskToDoService.TaskToDoRepository.GetTaskToDo(id);
 
             if (taskToDoToDelete == null)
             {
                 return NotFound($"TaskToDo with Id = {id} not found");
             }
 
-            return await _taskToDoService.DeleteTaskToDo(id);
+            return await _taskToDoService.TaskToDoRepository.DeleteTaskToDo(id);
         }
     }
 }
